rename numpy.py to numpy_wrapper.py to avoid lookup conflicts with python2.7
